feat: 카테고리 모달 내부의 종, 지역 리스트 컴포넌트로 분리
Dependency cycle 방지하기 위해 CategoryModalType 선언 위치 변경
